+++ tests/test_fgd.py
@@ -3,16 +3,16 @@ from typing import Any, Callable, Generator
 import copy
 import io
 
-import pytest
 from pytest_regressions.file_regression import FileRegressionFixture
+import pytest
 
 from srctools import Vec, fgd as fgd_mod
-from srctools.fgd import (
-    FGD, AutoVisgroup, EntityDef, EntityTypes, HelperExtAppliesTo, HelperHalfGridSnap,
-    HelperLine, HelperModel, HelperSize, HelperSphere, IODef, KVDef, Resource, Snippet,
-    UnknownHelper, ValueTypes,
-)
 from srctools.const import FileType
+from srctools.fgd import (
+    FGD, AutoVisgroup, EntityDef, EntityTypes, FGDParseError, HelperExtAppliesTo,
+    HelperHalfGridSnap, HelperLine, HelperModel, HelperSize, HelperSphere, IODef, KVDef,
+    Resource, Snippet, UnknownHelper, ValueTypes,
+)
 from srctools.filesys import VirtualFileSystem
 # noinspection PyProtectedMember
 from srctools.tokenizer import Cy_Tokenizer, Py_Tokenizer
@@ -99,6 +99,7 @@ appliesto(tag1, tag2, !tag3)
         2: "Third"
         "four": "Fourth"
         ]
+    ]
 """})
     fgd = FGD()
     fgd.parse_file(fsys, fsys['test.fgd'], eval_bases=False)
@@ -498,6 +499,44 @@ def test_snippet_io(py_c_token) -> None:
             ))
         )
     }
+
+
+PARSE_EOF = {
+    'entity.fgd': """\
+@PointClass = some_entity: "Some description" 
+[
+keyvalue1(string): "Name" : "default": "documentation"
+    """,
+    'entity_desc.fgd': """\
+@PointClass = some_entity: 
+    """,
+    'choices.fgd': """
+@PointClass = some_entity: "Some description" 
+[
+somechoices(choices): "Name" : 0 : "doc" = [
+"value1": "friendly"
+    """,
+    'spawnflags.fgd': """
+@PointClass = some_entity: "Some description" 
+[
+spawnflags(flags) = [
+1: "enable"
+2: "ignite"
+    """
+}
+
+
+@pytest.mark.parametrize('filename', PARSE_EOF)
+def test_parse_eof(py_c_token, filename: str) -> None:
+    """Test missing the ending bracket correctly causes errors."""
+    fgd = FGD()
+    fsys = VirtualFileSystem(PARSE_EOF)
+    with pytest.raises(FGDParseError) as err:
+        fgd.parse_file(fsys, fsys[filename])
+    assert 'ended unexpectedly' in err.value.mess
+    assert err.value.file == filename
+    # Don't particularly care where the line number is, just that it is set.
+    assert err.value.line_num is not None
 
 
 @pytest.mark.parametrize('custom_syntax', [False, True], ids=['vanilla', 'custom'])